+++ src/pykeen/nn/node_piece/representation.py
